requirements.txt
pyserial
pymodbus
twisted
mypy
